Created startpage and moved price adjustment
