+++ frontend/static/global.css
@@ -6,3 +6,12 @@
 .main{
     min-height: 80vh;
 }
+.center{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+.column{
+    display: flex;
+    flex-direction: column;
+}
